+++ packages/router/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 
+* The `active` interaction is now automatically included in a router.
 * A redirect can point to an external URL; the actual redirecting for external redirects is left to the user.
 * If the object returned by `route.response` has a property whose value is `undefined`, that property is left off of the `response`.
 * Change `emitRedirects` (default `true`) to `invisibleRedirects` (default `false`).